trigger delimiter only on found string
we can now choose to only remove content between
delimiters if a certain string is contained within that
string. Initially it was sought to do that using regular
expressions but as it seemed that was too slow.
A simple string search has to suffice for now but
the regex route may be explored further in the future
if need be.